Create README.md
Experiment with the different properties that affect text display, and various text formats to create effective and interesting page design. You can use for reference the constitution.html and style.css posted, which is just a plain black and white text version using some very basic CSS.

Be creative and come up with the formatting (text, background, spacing) to make this jump out and catch a viewer's attention.